Make memory usage reporting consistent

+++ src/KOKKOS/atom_vec_bond_kokkos.cpp
@@ -1148,9 +1148,9 @@ int AtomVecBondKokkos::write_data_hybrid(FILE *fp, double *buf)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-bigint AtomVecBondKokkos::memory_usage()
+double AtomVecBondKokkos::memory_usage()
 {
-  bigint bytes = 0;
+  double bytes = 0;
 
   if (atom->memcheck("tag")) bytes += memory->usage(tag,nmax);
   if (atom->memcheck("type")) bytes += memory->usage(type,nmax);
